Updates
Created a class Unit_Conveter, which contains 2 methods centimeter_to_millimeter and millimeter_to_centimeter to convert distance unit from cm to mm and mm to cm respectively.
Used vertical layout for the screen, added EditText of Numerical(decimal) type to enter a numerical value, RadioGroup which includes RadioButtons "From cm to mm" and
"From mm to cm"  which  allows to choose conversion from cm to mm or from mm to cm and a submit button "CONVERT" which calls the convert method and shows the result in EditText.
Added some background color to the app and changed font styles,font color of the texts on screen.